Bug fix newmsglist not checking for empty checkedlists

+++ msgs/newmsglist.php
@@ -23,14 +23,14 @@
 
 	$cid = $_GET['cid'];
 
-	if (isset($_POST['read'])) {
+	if (isset($_POST['read']) && count($_POST['checked'])>0) {
 		//DB $checklist = "'".implode("','",$_POST['checked'])."'";
 		$checklist = implode(',', array_map('intval', $_POST['checked']));
 		$query = "UPDATE imas_msgs SET isread=(isread|1) WHERE id IN ($checklist) AND (isread&1)=0";
 		//DB mysql_query($query) or die("Query failed : $query " . mysql_error());
 		$DBH->query($query);
 	}
-	if (isset($_POST['remove'])) {
+	if (isset($_POST['remove']) && count($_POST['checked'])>0) {
 		//DB $checklist = "'".implode("','",$_POST['checked'])."'";
 		$checklist = implode(',', array_map('intval', $_POST['checked']));
 		$query = "DELETE FROM imas_msgs WHERE id IN ($checklist) AND (isread&4)=4";
